ref(server-utils): Move server-only utils out of core
Moves a batch of server-only helpers out of `@sentry/core/server` into
`@sentry/server-utils`, continuing the effort to slim `@sentry/core` down to the
isomorphic surface. The moved symbols only ever depended on the public
`@sentry/core` API (plus the `@sentry/core/server` `filenameIsInApp` for anr), so
they relocate cleanly without expanding core's public surface.

Moved:
- `flushIfServerless` + `vercelWaitUntil` (used by the meta-framework SDKs)
- `trpcMiddleware`
- `callFrameToStackFrame` / `watchdogTimer` (the anr worker helpers)
- `loadModule` (split out of `utils/node`, `isNodeEnv` stays in core since core
  itself depends on it via `isBrowser`)

Every consumer already depends on `@sentry/server-utils`, so this only changes
the internal import path; the public SDK packages keep re-exporting these under
the same names, so there is no user-facing change.

`mcp-server` and the `integrations/http/*` subtree are intentionally left in core
for now: unlike the above, they import a number of non-public core internals
(`getSpanName`, `resolveDataCollectionOptions`, `getBreadcrumbLogLevel`,
`getDefaultExport`, `generatePropagationContext`, the internal `client`/`scope`
modules), so moving them would require promoting those to core's public API — a
separate decision.
