+++ packages/cloudflare/src/index.ts
@@ -106,7 +106,8 @@ export {
   withStreamedSpan,
   spanStreamingIntegration,
 } from '@sentry/core';
-export { instrumentPostgresJsSql, trpcMiddleware, wrapMcpServerWithSentry } from '@sentry/core/server';
+export { instrumentPostgresJsSql, wrapMcpServerWithSentry } from '@sentry/core/server';
+export { trpcMiddleware } from '@sentry/server-utils';
 
 export { withSentry } from './withSentry';
 export { defineCloudflareOptions } from './defineCloudflareOptions';
