+++ packages/core/src/server.ts
@@ -7,13 +7,9 @@
 export type { ServerRuntimeClientOptions } from './server-runtime-client';
 export { ServerRuntimeClient } from './server-runtime-client';
 export type { ServerRuntimeOptions } from './types/options';
-export { trpcMiddleware } from './trpc';
 export { wrapMcpServerWithSentry } from './integrations/mcp-server';
-export { isNodeEnv, loadModule } from './utils/node';
+export { isNodeEnv } from './utils/node';
 export { filenameIsInApp, node, nodeStackLineParser } from './utils/node-stack-trace';
-export { vercelWaitUntil } from './utils/vercelWaitUntil';
-export { flushIfServerless } from './utils/flushIfServerless';
-export { callFrameToStackFrame, watchdogTimer } from './utils/anr';
 export { safeUnref as _INTERNAL_safeUnref } from './utils/timer';
 /* oxlint-disable typescript/no-deprecated -- deprecated Express exports, kept until the next major */
 export { patchExpressModule } from './integrations/express/index';
