+++ packages/nuxt/src/runtime/hooks/wrapMiddlewareHandler.ts
@@ -11,7 +11,7 @@ import {
   type SpanAttributes,
   startSpan,
 } from '@sentry/core';
-import { flushIfServerless } from '@sentry/core/server';
+import { flushIfServerless } from '@sentry/server-utils';
 import type {
   _ResponseMiddleware as ResponseMiddleware,
   EventHandler,
