+++ packages/nuxt/test/runtime/hooks/wrapMiddlewareHandler.test.ts
@@ -1,5 +1,5 @@
 import * as SentryCore from '@sentry/core';
-import * as SentryCoreServer from '@sentry/core/server';
+import * as serverUtils from '@sentry/server-utils';
 import type { EventHandler, EventHandlerRequest, H3Event } from 'h3';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { wrapMiddlewareHandlerWithSentry } from '../../../src/runtime/hooks/wrapMiddlewareHandler';
@@ -17,13 +17,9 @@ vi.mock('@sentry/core', async importOriginal => {
   };
 });
 
-vi.mock('@sentry/core/server', async importOriginal => {
-  const mod = await importOriginal();
-  return {
-    ...(mod as any),
-    flushIfServerless: vi.fn(),
-  };
-});
+vi.mock('@sentry/server-utils', () => ({
+  flushIfServerless: vi.fn(),
+}));
 
 describe('wrapMiddlewareHandlerWithSentry', () => {
   const mockEvent: H3Event<EventHandlerRequest> = {
@@ -63,7 +59,7 @@ describe('wrapMiddlewareHandlerWithSentry', () => {
       }),
     });
     (SentryCore.httpHeadersToSpanAttributes as any).mockReturnValue({ 'http.request.header.user_agent': 'test-agent' });
-    (SentryCoreServer.flushIfServerless as any).mockResolvedValue(undefined);
+    (serverUtils.flushIfServerless as any).mockResolvedValue(undefined);
   });
 
   describe('function handler wrapping', () => {
