+++ packages/react-router/test/server/createSentryHandleError.test.ts
@@ -1,5 +1,5 @@
 import * as core from '@sentry/core';
-import * as coreServer from '@sentry/core/server';
+import * as serverUtils from '@sentry/server-utils';
 import type { ActionFunctionArgs, LoaderFunctionArgs } from 'react-router';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { createSentryHandleError } from '../../src/server/createSentryHandleError';
@@ -8,7 +8,7 @@ vi.mock('@sentry/core', () => ({
   captureException: vi.fn(),
 }));
 
-vi.mock('@sentry/core/server', () => ({
+vi.mock('@sentry/server-utils', () => ({
   flushIfServerless: vi.fn().mockResolvedValue(undefined),
 }));
 
@@ -19,7 +19,7 @@ const mechanism = {
 
 describe('createSentryHandleError', () => {
   const mockCaptureException = vi.mocked(core.captureException);
-  const mockFlushIfServerless = vi.mocked(coreServer.flushIfServerless);
+  const mockFlushIfServerless = vi.mocked(serverUtils.flushIfServerless);
   const mockConsoleError = vi.spyOn(console, 'error').mockImplementation(() => {});
 
   const mockError = new Error('Test error');
