+++ packages/server-utils/src/exports.ts
@@ -1,5 +1,10 @@
 // Shared exports not using diagnostics channels
 export { setHttpServerSpanRouteAttribute } from './utils/setHttpServerSpanRouteAttribute';
+export { flushIfServerless } from './utils/flushIfServerless';
+export { vercelWaitUntil } from './utils/vercelWaitUntil';
+export { loadModule } from './utils/loadModule';
+export { callFrameToStackFrame, watchdogTimer } from './utils/anr';
+export { trpcMiddleware } from './trpc';
 export { setAsyncLocalStorageAsyncContextStrategy } from './async-context';
 export { otlpIntegration, getOtlpTracesEndpoint } from './otlp';
 export * from './ai';
