+++ packages/server-utils/src/utils/loadModule.ts
@@ -1,0 +1,46 @@
+/**
+ * Requires a module which is protected against bundler minification.
+ *
+ * @param request The module path to resolve
+ */
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+function dynamicRequire(mod: any, request: string): any {
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+  return mod.require(request);
+}
+
+/**
+ * Helper for dynamically loading module that should work with linked dependencies.
+ * The problem is that we _should_ be using `require(require.resolve(moduleName, { paths: [cwd()] }))`
+ * However it's _not possible_ to do that with Webpack, as it has to know all the dependencies during
+ * build time. `require.resolve` is also not available in any other way, so we cannot create,
+ * a fake helper like we do with `dynamicRequire`.
+ *
+ * We always prefer to use local package, thus the value is not returned early from each `try/catch` block.
+ * That is to mimic the behavior of `require.resolve` exactly.
+ *
+ * @param moduleName module name to require
+ * @param existingModule module to use for requiring
+ * @returns possibly required module
+ */
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function loadModule<T>(moduleName: string, existingModule: any = module): T | undefined {
+  let mod: T | undefined;
+
+  try {
+    mod = dynamicRequire(existingModule, moduleName);
+  } catch {
+    // no-empty
+  }
+
+  if (!mod) {
+    try {
+      const { cwd } = dynamicRequire(existingModule, 'process');
+      mod = dynamicRequire(existingModule, `${cwd()}/node_modules/${moduleName}`) as T;
+    } catch {
+      // no-empty
+    }
+  }
+
+  return mod;
+}
