+++ packages/server-utils/test/utils/flushIfServerless.test.ts
@@ -1,8 +1,8 @@
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
-import * as flushModule from '../../../src/exports';
-import { flushIfServerless } from '../../../src/utils/flushIfServerless';
-import * as vercelWaitUntilModule from '../../../src/utils/vercelWaitUntil';
-import { GLOBAL_OBJ } from '../../../src/utils/worldwide';
+import * as SentryCore from '@sentry/core';
+import { flushIfServerless } from '../../src/utils/flushIfServerless';
+import * as vercelWaitUntilModule from '../../src/utils/vercelWaitUntil';
+import { GLOBAL_OBJ } from '@sentry/core';
 
 describe('flushIfServerless', () => {
   let originalProcess: typeof process;
@@ -17,7 +17,7 @@ describe('flushIfServerless', () => {
   });
 
   test('should bind context (preserve `this`) when calling waitUntil from the Cloudflare execution context', async () => {
-    const flushMock = vi.spyOn(flushModule, 'flush').mockResolvedValue(true);
+    const flushMock = vi.spyOn(SentryCore, 'flush').mockResolvedValue(true);
 
     // Mock Cloudflare context with `waitUntil` (which should be called if `this` is bound correctly)
     const mockCloudflareCtx = {
@@ -38,7 +38,7 @@ describe('flushIfServerless', () => {
   });
 
   test('should use cloudflare waitUntil when valid cloudflare context is provided', async () => {
-    const flushMock = vi.spyOn(flushModule, 'flush').mockResolvedValue(true);
+    const flushMock = vi.spyOn(SentryCore, 'flush').mockResolvedValue(true);
     const mockCloudflareCtx = {
       waitUntil: vi.fn(),
     };
@@ -50,7 +50,7 @@ describe('flushIfServerless', () => {
   });
 
   test('should use cloudflare waitUntil when Cloudflare `waitUntil` is provided', async () => {
-    const flushMock = vi.spyOn(flushModule, 'flush').mockResolvedValue(true);
+    const flushMock = vi.spyOn(SentryCore, 'flush').mockResolvedValue(true);
     const mockCloudflareCtx = {
       waitUntil: vi.fn(),
     };
@@ -62,7 +62,7 @@ describe('flushIfServerless', () => {
   });
 
   test('should ignore cloudflare context when waitUntil is not a function (and use Vercel waitUntil instead)', async () => {
-    const flushMock = vi.spyOn(flushModule, 'flush').mockResolvedValue(true);
+    const flushMock = vi.spyOn(SentryCore, 'flush').mockResolvedValue(true);
     const vercelWaitUntilSpy = vi.spyOn(vercelWaitUntilModule, 'vercelWaitUntil').mockImplementation(() => {});
 
     // Mock Vercel environment
@@ -81,7 +81,7 @@ describe('flushIfServerless', () => {
   });
 
   test('should handle multiple serverless environment variables simultaneously', async () => {
-    const flushMock = vi.spyOn(flushModule, 'flush').mockResolvedValue(true);
+    const flushMock = vi.spyOn(SentryCore, 'flush').mockResolvedValue(true);
 
     global.process = {
       ...originalProcess,
@@ -100,7 +100,7 @@ describe('flushIfServerless', () => {
   });
 
   test('should use default timeout when not specified', async () => {
-    const flushMock = vi.spyOn(flushModule, 'flush').mockResolvedValue(true);
+    const flushMock = vi.spyOn(SentryCore, 'flush').mockResolvedValue(true);
     const mockCloudflareCtx = {
       waitUntil: vi.fn(),
     };
@@ -111,7 +111,7 @@ describe('flushIfServerless', () => {
   });
 
   test('should handle zero timeout value', async () => {
-    const flushMock = vi.spyOn(flushModule, 'flush').mockResolvedValue(true);
+    const flushMock = vi.spyOn(SentryCore, 'flush').mockResolvedValue(true);
 
     global.process = {
       ...originalProcess,
