+++ packages/vercel-edge/src/index.ts
@@ -101,7 +101,8 @@ export {
   withStreamedSpan,
   spanStreamingIntegration,
 } from '@sentry/core';
-export { trpcMiddleware, wrapMcpServerWithSentry } from '@sentry/core/server';
+export { wrapMcpServerWithSentry } from '@sentry/core/server';
+export { trpcMiddleware } from '@sentry/server-utils';
 export {
   otlpIntegration,
   getOtlpTracesEndpoint,
